+++ Js2IL.Tests/CommonJS/JavaScript/CommonJS_Export_Class_Lib.js
@@ -1,0 +1,14 @@
+// Library module that exports an object instance with methods
+
+class Calculator {
+    add(a, b) {
+        return a + b;
+    }
+    
+    multiply(a, b) {
+        return a * b;
+    }
+}
+
+// Export an instance of the class
+module.exports = new Calculator();