+++ Js2IL.Tests/CommonJS/JavaScript/CommonJS_Export_Function_Lib.js
@@ -1,0 +1,7 @@
+// Library module that exports a single function
+
+function greet(name) {
+    return "Hello, " + name + "!";
+}
+
+module.exports = greet;